moved parenthesis handler and strategies to its own dir, updated imports
