+++ ingredient_slicer/parenthesis/equivalence_parenthesis_strategy.py
@@ -5,7 +5,7 @@
 
 from ingredient_slicer import _constants
 from ingredient_slicer import _utils
-from ingredient_slicer.parser._parenthesis_handler import ParenthesisHandler
+from ingredient_slicer.parenthesis.parenthesis_handler import ParenthesisHandler
 from ingredient_slicer.models.quantity_unit_data import QuantityUnitData
 
 class EquivalenceParenthesisHandler(ParenthesisHandler):
