+++ ingredient_slicer/parenthesis/quantity_unit_only_parenthesis_strategy.py
@@ -1,6 +1,6 @@
 from ingredient_slicer import _constants 
 from ingredient_slicer import _utils
-from ingredient_slicer.parser._parenthesis_handler import ParenthesisHandler
+from ingredient_slicer.parenthesis.parenthesis_handler import ParenthesisHandler
 from ingredient_slicer.models.quantity_unit_data import QuantityUnitData
 
 class QuantityUnitOnlyParenthesisHandler(ParenthesisHandler):
