+++ ingredient_slicer/parser/ingredient_parser.py
@@ -11,10 +11,10 @@ from ingredient_slicer import _constants
 
 from ingredient_slicer.standardizer._ingredient_standardizer import IngredientStandardizer 
 
-from ingredient_slicer.parser._parenthesis_handler import ParenthesisHandler
-from ingredient_slicer.parser._quantity_only_parenthesis_strategy import QuantityOnlyParenthesisHandler 
-from ingredient_slicer.parser._equivalence_parenthesis_strategy import EquivalenceParenthesisHandler 
-from ingredient_slicer.parser._quantity_unit_only_parenthesis_strategy import QuantityUnitOnlyParenthesisHandler
+from ingredient_slicer.parenthesis.parenthesis_handler import ParenthesisHandler
+from ingredient_slicer.parenthesis.quantity_only_parenthesis_strategy import QuantityOnlyParenthesisHandler 
+from ingredient_slicer.parenthesis.equivalence_parenthesis_strategy import EquivalenceParenthesisHandler 
+from ingredient_slicer.parenthesis.quantity_unit_only_parenthesis_strategy import QuantityUnitOnlyParenthesisHandler
 
 from ingredient_slicer.models.quantity_unit_data import QuantityUnitData
 from ingredient_slicer.models.parsed_ingredient_data import ParsedIngredientData
